Replace rust-crypto with sha2 crate
The RustSec security advisory database lists rust-crypto as
unmaintained. As such it is recommended that the crate is replaced with
a maintained alternative. This patch replaces the rust-crypto crate with
the sha2 crate and updates the affected hashing logic.

The patch also contains a change to a test that fails on windows.