Replace rust-crypto with sha2 crate (#10)
* Fix failing tests on windows

Windows uses backslashes as path separator while some tests assumed
forward slashes. This patch changes those tests to use a
platform-dependent separator.

* Replace rust-crypto with sha2 crate

The RustSec security advisory database lists rust-crypto as
unmaintained. As such it is recommended that the crate is replaced with
a maintained alternative. This patch replaces the rust-crypto crate with
the sha2 crate and updates the affected hashing logic.

The patch also contains a change to a test that fails on windows.

* Patch httmock to fix a bug

Check https://github.com/alexliesenfeld/httpmock/pull/4 for more details.

Co-authored-by: Evan Pete Walsh <epwalsh10@gmail.com>