refactor crate to use async/tokio over blocking calls
The current use of `reqwest::blocking` panics when used in an async environment. This commit keeps all functionality as is but uses the Tokio runtime to make it work in an async environment. This means in turn that the current crate does not work in a sync environment. Omitted are any updates regarding documentation.

```
#[tokio::main]
async fn main() {
    let cache_dir = dirs::cache_dir().unwrap();
    let remote = "https://...";
    let local_path = "my/download/path";

    let cache = Cache::builder()
        .dir(cache_dir)
        .progress_bar(None)
        .build()
        .unwrap();

    let path = self
        .cache
        .cached_path_with_options(&remote, &Options::default().subdir(&local_path))
        .await
        .unwrap();

    println!("{:?}", path);
}
```